// GC-Mobile-Web-Application/UnitedAdminPortal/src/app/models/promotion.ts
export class Promotion{
  promoID: string;
  eventID: string;
  userID: string;
  offerTitle: string;
  promoImageUrl: string;
  destinationLink: string;
  promoType: string;
  mainStyle: string;
  relatedStyle1: string;
  relatedStyle2: string;
  relatedStyle3: string;
  promoCity: string;
  viewCounter: number;
}